Add StoryContainer, separate MainLayout styles
This moves the horizontal padding from MainLayout into the new
StoryContainer so that Gallery components within MainLayout can be the
full width of the screen, and the text within StoryContainer makes room
for the main menu if present.

This also updates the default showMenu prop of MainLayout to be true
since that's how it's actually working out. Less props to pass.