Add theme approximation for storybook
This adds the CssVariablesTheme AddOn for Storybook, and implements
theming slightly different than we do in the actual app: these new
files load CSS variables on the :root object when it loads the styles;
in the app we're setting a className on MainLayout for defining color
variables.

Because of this, the MainLayout story is a little wonky because there
can be conflicting settings between the storybook add on and the
isDarkMode state.