Finalize Gallery component styles and animation
This incorporates the example at https://codesandbox.io/s/framer-motion-image-gallery-pqvx3
to finish setting up the Gallery. The component now accepts a slides
array of strings, expected to be urls for each image. This prop will
likely become an object in a future commit to also handle alt text and
support youtube videos.

The stories showcase that we may need to handle thumbnail styles at
smaller screen sizes, especially if there are a lot of images.

Also contains tweaks to related styles:
- removes unnecessary explicit width on ExpandableBlock
- specified z-index for menuButton so it remains on top of Gallery