Add basic scroll animation from MainMenu links
This sets up anchor links for each MainLayout section in the App
component... which is making me want to rethink this architecture. I
think at this point the functionality of MainLayout can be incorporated
into App, and maybe we'll want some kind of component to handle common
styling for each section that can be navigated to? Unclear, but will
address next.

For this commit, each section gets a good old fashioned ID attribute,
and then the MainMenu utilizes vanilla JS's scrollIntoView to animate
the scroll to the right anchor link.