Add Caption component, rework Gallery styles
This reworks the Gallery styles to incorporate a caption component
under the 16:9 formatted space for media. Because Framer Motion wants
to animate on the x axis, slides need to be absolutely positioned,
meaning the container needs an explicit height or everything will
collapse.

Instead of trying to animate a variable height based on caption text,
this sets enough vertical space for three lines of caption text, and
leaves that same amount of space for every slide.

Also there are a lot of similarities between SingleMedia and Gallery
and it's probably time to combine these so I don't have repeated code.