Track current chapter for routing and header text
This uses the FIREBASE_DOC_ORDER constant for determining the index of
the currently-viewed chapter to update both routing and dynamic header
text.

The react-intersection-observer library determines if a Section
component is in view while scrolling. If the Section is in view, it
will be set as the currentIndex.

The currentIndex is used to determine the chapterTitle prop for Header
and when it is updated we also update the url routing to match.