Add Player component for lyrics + media
This adds a modal for displaying Player on click of PlayerButton, using
the href from any a tag in a markdown file as a mediaId for later
grabbing the correct data for each song.

Also moves the closeX icon from MainMenu to icons since we're using the
same one to close the Player modal.